# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DifficultyLevel": {
        "enum": [
          "EASY",
          "MEDIUM",
          "HARD"
        ],
        "title": "DifficultyLevel",
        "type": "string"
      },
...
      "ExamCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "default": 60,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "is_published": {
            "default": false,
            "title": "Is Published",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "title"
        ],
        "title": "ExamCreate",
        "type": "object"
      },
      "ExamDetailResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creator_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "default": 60,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_published": {
            "default": false,
            "title": "Is Published",
            "type": "boolean"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "title": "Questions",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "title",
          "id",
          "creator_id",
          "created_at",
          "questions"
        ],
        "title": "ExamDetailResponse",
        "type": "object"
      },
      "ExamResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creator_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "default": 60,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_published": {
            "default": false,
            "title": "Is Published",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "title",
          "id",
          "creator_id",
          "created_at"
        ],
        "title": "ExamResponse",
        "type": "object"
      },
      "ExamUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "default": 60,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "is_published": {
            "default": false,
            "title": "Is Published",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "title"
        ],
        "title": "ExamUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionBase": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "is_correct": {
            "title": "Is Correct",
            "type": "boolean"
          }
        },
        "required": [
          "content",
          "is_correct"
        ],
        "title": "OptionBase",
        "type": "object"
      },
      "OptionResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_correct": {
            "title": "Is Correct",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "content",
          "is_correct"
        ],
        "title": "OptionResponse",
        "type": "object"
      },
...
      "Page_ExamResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1,
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "minimum": 0,
            "title": "Pages",
            "type": "integer"
          },
          "size": {
            "minimum": 1,
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[ExamResponse]",
        "type": "object"
      },
...
      "Page_StudentExamListResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentExamListResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1,
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "minimum": 0,
            "title": "Pages",
            "type": "integer"
          },
          "size": {
            "minimum": 1,
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[StudentExamListResponse]",
        "type": "object"
      },
...
      "QuestionBase": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel",
            "default": "MEDIUM"
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "options": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OptionBase"
            },
            "title": "Options",
            "type": "array"
          },
          "points": {
            "default": 1,
            "title": "Points",
            "type": "number"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType",
            "default": "MULTIPLE_CHOICE"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "content"
        ],
        "title": "QuestionBase",
        "type": "object"
      },
      "QuestionBulkCreate": {
        "properties": {
          "question_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Question Ids",
            "type": "array"
          }
        },
        "required": [
          "question_ids"
        ],
        "title": "QuestionBulkCreate",
        "type": "object"
      },
      "QuestionResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_ai_generated": {
            "title": "Is Ai Generated",
            "type": "boolean"
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionResponse"
            },
            "title": "Options",
            "type": "array"
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "id",
          "content",
          "is_ai_generated",
          "points",
          "question_type",
          "difficulty",
          "options"
        ],
        "title": "QuestionResponse",
        "type": "object"
      },
...
      "QuestionType": {
        "enum": [
          "SINGLE_CHOICE",
          "MULTIPLE_CHOICE",
          "MATCHING",
          "FILL_IN_BLANK"
        ],
        "title": "QuestionType",
        "type": "string"
      },
...
      "StudentExamListResponse": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "submission_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Score"
          }
        },
        "required": [
          "id",
          "title",
          "duration_minutes"
        ],
        "title": "StudentExamListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "AI Quiz System",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/exams": {
      "get": {
        "operationId": "read_exams_exams_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "topic_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          },
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExamResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Exams",
        "tags": [
          "exams"
        ]
      },
      "post": {
        "operationId": "create_exam_exams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/exams/{exam_id}": {
      "delete": {
        "description": "Delete an exam. (Requires creator or admin authentication)",
        "operationId": "delete_exam_exams__exam_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Exam",
        "tags": [
          "exams"
        ]
      },
      "get": {
        "operationId": "get_exam_exams__exam_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Exam",
        "tags": [
          "exams"
        ]
      },
      "put": {
        "operationId": "update_exam_exams__exam_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/exams/{exam_id}/questions": {
      "post": {
        "operationId": "create_question_in_exam_exams__exam_id__questions_post",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Question In Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/exams/{exam_id}/questions/bulk": {
      "post": {
        "operationId": "add_questions_to_exam_exams__exam_id__questions_bulk_post",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBulkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Questions To Exam Exams  Exam Id  Questions Bulk Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Questions To Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/exams/{exam_id}/restore": {
      "post": {
        "description": "Restore a soft-deleted exam within the 30-day retention window.\n(Requires creator or admin authentication)",
        "operationId": "restore_exam_exams__exam_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Restore Exam",
        "tags": [
          "exams"
        ]
      }
    },
...
    "/student/exams": {
      "get": {
        "operationId": "get_available_exams_student_exams_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StudentExamListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Available Exams",
        "tags": [
          "student"
        ]
      }
    },
...
  }
}